HTML setup, title, and ingredient table
Transformation converts the recipies.xml to include
the html tags needed, a title for each recipe, and
the ingredient table.

Ingredient table can have super-ingredients. This is
shown by sectioning off the with a table header
and listing the ingredients needed to make the super
ingredient.

Also included are the pluralization of the ingredient
unit when needed (e.g 2 cups, 1 teaspoon).